pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "alembic==1.17.2",
    "anthropic==0.73.0",
    "argon2-cffi>=25.1.0",
    "black>=26.3.1",
    "bs4>=0.0.2",
    "chromadb==1.3.5",
    "dotenv>=0.9.9",
    "fastapi==0.124.2",
    "fastmcp==2.13.1",
    "google-genai==1.50.1",
    "langchain>=0.3.28",
    "langchain-community>=0.3.31",
    "langchain-core<0.4",
    "langchain-ollama>=0.3.10",
    "langchain-openai>=0.3.35",
    "langchain-postgres>=0.0.16",
    "logfire==4.14.2",
    "openai==2.9.0",
    "passlib>=1.7.4",
    "pgvector<0.4",
    "pre-commit==4.5.0",
    "psycopg[binary]==3.3.2",
    "pydantic-settings>=2.13.1",
    "pymupdf>=1.26.7",
    "pytest==9.0.1",
    "pytest-asyncio>=0.23.0",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.45",
    "streamlit==1.51.0",
    "uvicorn>=0.30.0",
]
